% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19Data.R
\name{Covid19Data}
\alias{Covid19Data}
\title{Covid19Data}
\usage{
Covid19Data()
}
\value{
a dataframe with the daily worlwide data of covid19 per country, with total and new confirmed cases and  death
}
\description{
function to read "live" covid19 data from Johns Hopkins University data repository.

Note: This data is based on the information each country reported every day of it's total confirmed cases or death. Some countries on some dates reported number of it's cases less than the day before, so the new cases for them on that date will be reported negative in the data set.
}
\examples{
\dontrun{
newdata <- Covid19Data()
}
}
