% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19onMap.R
\name{Covid19onMap}
\alias{Covid19onMap}
\title{Covid19onMap}
\usage{
Covid19onMap(data = local_data, date, type)
}
\arguments{
\item{data}{a dataframe that contains covid19 data which can be reached by function Covid19Data(). defult is local_data which is the old version of covid19 data, that is stored in the package. some one can use this defult if he/she can't reach the net}

\item{date}{a string that indicate a date, possible formats are: "y-m-d" or "y/m/d"}

\item{type}{a string that indicate type of reported cases, possible values are: "confirmed" and "death"}
}
\value{
a world map that the color of every country shows number of new cases of confirmed or death from covid19 on the given date
}
\description{
function to show number of new cases of confirmed or death from covid19 of every country on the world map on a specific date
}
\examples{
\dontrun{
newdata <- Covid19Data()
Covid19onMap(newdata, "2020-05-24", "confirmed")
Covid19onMap(newdata, "2020/05/24", "death")
}
}
