% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19Plot.R
\name{Covid19Plot}
\alias{Covid19Plot}
\title{Covid19Plot}
\usage{
Covid19Plot(
  data = local_data,
  country,
  start_date = "2020-01-22",
  end_date = Sys.Date()
)
}
\arguments{
\item{data}{a dataframe that contains covid19 data which can be reached by function Covid19Data(). defult is local_data which is the old version of covid19 data, that is stored in the package. some one can use this defult if he/she can't reach the net}

\item{country}{a string that indicate name of a country or a vector of strings that indicate name of some countries, available names are stored in the variable named 'country_names'}

\item{start_date}{a string that indicate the beginning of the interval of date, possible formats are: "y-m-d" or "y/m/d"}

\item{end_date}{a string that indicate the end of the interval of date, possible formats are: "y-m-d" or "y/m/d"}
}
\value{
a plot of number of new cases of confirmed and death from covid19 as a function of time for the given country in the given time interval
}
\description{
function to plot number of new cases of confirmed and death from covid19 as a function of time for a specific country in a time interval
}
\examples{
\dontrun{
newdata <- Covid19Data()
Covid19Plot(newdata,"Iran")
Covid19Plot(newdata,"Iran",start_date="2020-01-22",end_date = Sys.Date())
Covid19Plot(newdata,country=c("Iran","india"),start_date="2020/02/22",end_date = "2020/05/15")
}
}
