% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19Data.R
\name{Covid19Data}
\alias{Covid19Data}
\title{Covid19Data}
\usage{
Covid19Data()
}
\value{
a dataframe with the daily worlwide data of covid19 per country with total and new cases of confirmed and death
}
\description{
function to read "live" covid19 data from Johns Hopkins University data repository
}
\examples{
\dontrun{
newdata <- Covid19Data()
}
}
