% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19Plot.R
\name{Covid19Plot}
\alias{Covid19Plot}
\title{Covid19Plot}
\usage{
Covid19Plot(
  data = local_data,
  country,
  start_date = "2020-01-22",
  end_date = Sys.Date()
)
}
\arguments{
\item{data}{a dataframe that contains covid19 data which can be reached by function Covid19Data(). Default is local_data which is the offline available version of covid19 data stored in the package. Someone can use this default if he/she can't reach the net}

\item{country}{a string that indicates name of a country or a vector of strings that indicates name of some countries. available names are stored in the variable named 'country_names'}

\item{start_date}{a string that indicates the beginning of the interval, possible formats are: "yyyy-m-d" or "yyyy/m/d"}

\item{end_date}{a string that indicates the end of the interval, possible formats are: "yyyy-m-d" or "yyyy/m/d"}
}
\value{
a plot of number of daily confirmed cases and death from covid19 as a function of time for the given country in the given time interval
}
\description{
function to plot number of daily confirmed cases and death from covid19 as a function of time for a given country in a given time interval
}
\examples{
\dontrun{
newdata <- Covid19Data()
Covid19Plot(newdata,"Iran")
Covid19Plot(newdata,"Iran",start_date="2020-01-22",end_date = Sys.Date())
Covid19Plot(newdata,country=c("Iran","india"),start_date="2020/02/22",end_date = "2020/05/15")
}
}
